Add files via upload
Try to support MIDI 3d in textmidi.  Protect against overwriting the input fie in miditext.  Add policy for Running Status:
per spec or never.  This made it possible to compare true to about 88% of MIDI files in a round-trip
from midi to text to midi, sometimes using per-spec running status, and sometimes not using it at all.
Use more C++20: optional<> used for the bidirectional maps to fail with nothing.