Add files via upload
Rename track scrambling to arrangements.  Improve documentation of arrangements.
Change arrangement "None" to "Identity" and document.  Remove the linefeed from 
definition of LINE.
Introduce a Heap's Algorithm permutation to Arrangements.
Redefine trackscrambling as arrangements so that it is easier to explain.
String escape and quotes handling moved out of textmidi and miditext.