Add files via upload
Compile with -std=c++23
Fix guitar range in textmidiformedit.py and textmidicgm; it was an octave high, yes I know the guitar's range, it was a clerical error from going too quickly.
textmidiformedit.py: transpose spinbox didn't update on file loads.
textmidicgm: a lower-numbered track couldn't follow a higher-numbered track; this was not the intention.  Track 0 should be able to follow track 1.
The symptom was that no tracks were generated.  The cause was using "i" index instead of "tr", from the topological sort of tracks.  "i" has been deleted.